[Refactor] Reorganize tests
* Split a big test into multiple small ones.

* Created `asio` directory and moved all asio involved tests to there.
