+++ scripts/test/runtests.sh
@@ -3,20 +3,25 @@ set -e
 
 ./tests/buffer_test --abort-on-failure
 ./tests/serialization_test --abort-on-failure
-if [ -f ./tests/timer_test ]; then
-    ./tests/timer_test --abort-on-failure
-fi
 ./tests/strfmt_test --abort-on-failure
 ./tests/stat_mgr_test --abort-on-failure
 ./tests/raft_server_test --abort-on-failure
+./tests/snapshot_test --abort-on-failure
+./tests/leader_election_test --abort-on-failure
 ./tests/new_joiner_test --abort-on-failure
 ./tests/failure_test --abort-on-failure
-if [ -f ./tests/asio_service_test ]; then
-    ./tests/asio_service_test --abort-on-failure
-fi
-if [ -f ./tests/asio_service_stream_test ]; then
-    ./tests/asio_service_stream_test --abort-on-failure
-fi
-if [ -f ./tests/stream_functional_test ]; then
-    ./tests/stream_functional_test --abort-on-failure
-fi
+
+ASIO_TESTS=(
+    "./tests/timer_test"
+    "./tests/asio_service_test"
+    "./tests/req_resp_meta_test"
+    "./tests/custom_quorum_test"
+    "./tests/stream_transport_layer_test"
+    "./tests/raft_stream_mode_test"
+)
+
+for test in "${ASIO_TESTS[@]}"; do
+    if [ -f $test ]; then
+        $test --abort-on-failure
+    fi
+done
