+++ tests/asio/asio_service_test.cxx
@@ -1,5 +1,5 @@
 /************************************************************************
-Copyright 2017-2019 eBay Inc.
+Copyright 2017-present eBay Inc.
 Author/Developer(s): Jung-Sang Ahn
 
 Licensed under the Apache License, Version 2.0 (the "License");
@@ -19,6 +19,7 @@ limitations under the License.
 #include "debugging_options.hxx"
 #include "in_memory_log_store.hxx"
 #include "raft_package_asio.hxx"
+#include "asio_test_common.hxx"
 
 #include "event_awaiter.hxx"
 #include "test_common.h"
@@ -41,53 +42,7 @@ using namespace raft_functional_common;
 
 static bool flag_bg_snapshot_io = false;
 
-std::ostream& operator<<(std::ostream& os, raft_server::PrioritySetResult res) {
-    if (res == raft_server::PrioritySetResult::SET)
-        os << "SET";
-    else if (res == raft_server::PrioritySetResult::IGNORED)
-        os << "IGNORED";
-    else
-        os << "BROADCAST";
-    return os;
-}
-
 namespace asio_service_test {
-
-int launch_servers(const std::vector<RaftAsioPkg*>& pkgs,
-                   bool enable_ssl,
-                   bool use_global_asio = false,
-                   bool use_bg_snapshot_io = true,
-                   const raft_server::init_options & opt = raft_server::init_options())
-{
-    size_t num_srvs = pkgs.size();
-    CHK_GT(num_srvs, 0);
-
-    for (auto& entry: pkgs) {
-        RaftAsioPkg* pp = entry;
-        pp->initServer(enable_ssl, use_global_asio, use_bg_snapshot_io, opt);
-    }
-    // Wait longer than upper timeout.
-    TestSuite::sleep_sec(1);
-    return 0;
-}
-
-int make_group(const std::vector<RaftAsioPkg*>& pkgs) {
-    size_t num_srvs = pkgs.size();
-    CHK_GT(num_srvs, 0);
-
-    RaftAsioPkg* leader = pkgs[0];
-
-    for (size_t ii = 1; ii < num_srvs; ++ii) {
-        RaftAsioPkg* ff = pkgs[ii];
-
-        // Add to leader.
-        leader->raftServer->add_srv( *(ff->getTestMgr()->get_srv_config()) );
-
-        // Wait longer than upper timeout.
-        TestSuite::sleep_sec(1);
-    }
-    return 0;
-}
 
 int make_group_test() {
     reset_log_files();
@@ -338,557 +293,6 @@ int ssl_test() {
     return 0;
 }
 
-static bool dbg_print_ctx = false;
-static std::unordered_map<std::string, std::string> req_map;
-static std::unordered_map<std::string, std::string> resp_map;
-static std::mutex req_map_lock;
-static std::mutex resp_map_lock;
-
-std::string test_write_req_meta( std::atomic<size_t>* count,
-                                 const asio_service::meta_cb_params& params )
-{
-    static std::mutex lock;
-    std::lock_guard<std::mutex> l(lock);
-
-    char key[256];
-    snprintf(key, 256, "%2d, %2d -> %2d, %4zu",
-             params.msg_type_, params.src_id_, params.dst_id_,
-             (size_t)params.log_idx_);
-
-    std::string value = "req_" + std::to_string( std::rand() );
-    {   std::lock_guard<std::mutex> l(req_map_lock);
-        req_map[key] = value;
-    }
-
-    if (dbg_print_ctx) {
-        _msg("%10s %s %20s\n", "write req", key, value.c_str());
-    }
-
-    // `req_` should be given, while `resp_` should be null.
-    auto chk_req_resp = [&]() {
-        CHK_NONNULL(params.req_);
-        CHK_NULL(params.resp_);
-        return 0;
-    };
-    if (chk_req_resp() != 0) return std::string();
-
-    if (count) (*count)++;
-    return value;
-}
-
-bool test_read_req_meta( std::atomic<size_t>* count,
-                         const asio_service::meta_cb_params& params,
-                         const std::string& meta )
-{
-    static std::mutex lock;
-
-    std::lock_guard<std::mutex> l(lock);
-
-    char key[256];
-    snprintf(key, 256, "%2d, %2d -> %2d, %4zu",
-             params.msg_type_, params.src_id_, params.dst_id_,
-             (size_t)params.log_idx_);
-
-    if (dbg_print_ctx) {
-        _msg("%10s %s %20s\n", "read req", key, meta.c_str());
-    }
-
-    // `req_` should be given, while `resp_` should be null.
-    auto chk_req_resp = [&]() {
-        CHK_NONNULL(params.req_);
-        CHK_NULL(params.resp_);
-        return 0;
-    };
-    if (chk_req_resp() != 0) return false;
-
-    std::string META;
-    {   std::lock_guard<std::mutex> l(req_map_lock);
-        META = req_map[key];
-    }
-    if (META != meta) {
-        CHK_EQ( META, meta );
-        return false;
-    }
-    if (count) (*count)++;
-    return true;
-}
-
-std::string test_write_resp_meta( std::atomic<size_t>* count,
-                                  const asio_service::meta_cb_params& params )
-{
-    static std::mutex lock;
-    std::lock_guard<std::mutex> l(lock);
-
-    char key[256];
-    snprintf(key, 256, "%2d, %2d -> %2d, %4zu",
-             params.msg_type_, params.src_id_, params.dst_id_,
-             (size_t)params.log_idx_);
-
-    std::string value = "resp_" + std::to_string( std::rand() );
-    {   std::lock_guard<std::mutex> l(resp_map_lock);
-        resp_map[key] = value;
-    }
-
-    if (dbg_print_ctx) {
-        _msg("%10s %s %20s\n", "write resp", key, value.c_str());
-    }
-
-    // `req_` and `resp_` should be given.
-    auto chk_req_resp = [&]() {
-        CHK_NONNULL(params.req_);
-        CHK_NONNULL(params.resp_);
-        return 0;
-    };
-    if (chk_req_resp() != 0) return std::string();
-
-    if (count) (*count)++;
-    return value;
-}
-
-bool test_read_resp_meta( std::atomic<size_t>* count,
-                          const asio_service::meta_cb_params& params,
-                          const std::string& meta )
-{
-    static std::mutex lock;
-    std::lock_guard<std::mutex> l(lock);
-
-    char key[256];
-    snprintf(key, 256, "%2d, %2d -> %2d, %4zu",
-             params.msg_type_, params.src_id_, params.dst_id_,
-             (size_t)params.log_idx_);
-
-    if (dbg_print_ctx) {
-        _msg("%10s %s %20s\n", "read resp", key, meta.c_str());
-    }
-
-    // `req_` and `resp_` should be given.
-    auto chk_req_resp = [&]() {
-        CHK_NONNULL(params.req_);
-        CHK_NONNULL(params.resp_);
-        return 0;
-    };
-    if (chk_req_resp() != 0) return false;
-
-    std::string META;
-    {   std::lock_guard<std::mutex> l(resp_map_lock);
-        META = resp_map[key];
-    }
-    if (META != meta) {
-        CHK_EQ( META, meta );
-        return false;
-    }
-    if (count) (*count)++;
-    return true;
-}
-
-int message_meta_test(bool crc_on_entire_message) {
-    reset_log_files();
-
-    std::string s1_addr = "127.0.0.1:20010";
-    std::string s2_addr = "127.0.0.1:20020";
-    std::string s3_addr = "127.0.0.1:20030";
-
-    RaftAsioPkg s1(1, s1_addr);
-    RaftAsioPkg s2(2, s2_addr);
-    RaftAsioPkg s3(3, s3_addr);
-    std::vector<RaftAsioPkg*> pkgs = {&s1, &s2, &s3};
-
-    std::atomic<size_t> read_req_cb_count(0);
-    std::atomic<size_t> write_req_cb_count(0);
-    std::atomic<size_t> read_resp_cb_count(0);
-    std::atomic<size_t> write_resp_cb_count(0);
-
-    _msg("launching asio-raft servers with meta callback\n");
-    for (RaftAsioPkg* rr: pkgs) {
-        rr->setMetaCallback
-            ( std::bind( test_read_req_meta,
-                         &read_req_cb_count,
-                         std::placeholders::_1,
-                         std::placeholders::_2 ),
-              std::bind( test_write_req_meta,
-                         &write_req_cb_count,
-                         std::placeholders::_1 ),
-              std::bind( test_read_resp_meta,
-                         &read_resp_cb_count,
-                         std::placeholders::_1,
-                         std::placeholders::_2 ),
-              std::bind( test_write_resp_meta,
-                         &write_resp_cb_count,
-                         std::placeholders::_1 ),
-              true );
-
-        rr->setCrcOnEntireMessage(crc_on_entire_message);
-    }
-    CHK_Z( launch_servers(pkgs, false) );
-
-    _msg("organizing raft group\n");
-    CHK_Z( make_group(pkgs) );
-
-    CHK_TRUE( s1.raftServer->is_leader() );
-    CHK_EQ(1, s1.raftServer->get_leader());
-    CHK_EQ(1, s2.raftServer->get_leader());
-    CHK_EQ(1, s3.raftServer->get_leader());
-
-    TestSuite::sleep_sec(1, "wait for Raft group ready");
-
-    for (size_t ii=0; ii<10; ++ii) {
-        std::string msg_str = std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(sizeof(uint32_t) + msg_str.size());
-        buffer_serializer bs(msg);
-        bs.put_str(msg_str);
-        s1.raftServer->append_entries( {msg} );
-    }
-
-    TestSuite::sleep_sec(1, "wait for replication");
-
-    // Callback functions for meta should have been called.
-    CHK_GT(read_req_cb_count.load(), 0);
-    CHK_GT(write_req_cb_count.load(), 0);
-    CHK_GT(read_resp_cb_count.load(), 0);
-    CHK_GT(write_resp_cb_count.load(), 0);
-    _msg( "read req callback %zu, write req callback %zu\n",
-          read_req_cb_count.load(), write_req_cb_count.load() );
-    _msg( "read resp callback %zu, write resp callback %zu\n",
-          read_resp_cb_count.load(), write_resp_cb_count.load() );
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-    TestSuite::sleep_sec(1, "shutting down");
-
-    SimpleLogger::shutdown();
-    return 0;
-}
-
-bool test_read_meta_random_denial( std::atomic<bool>* start_denial,
-                                   const asio_service::meta_cb_params& params,
-                                   const std::string& meta )
-{
-    if ( !(start_denial->load()) ) return true;
-
-    int r = std::rand();
-    if (r % 25 == 0) return false;
-    return true;
-}
-
-int message_meta_random_denial_test() {
-    reset_log_files();
-
-    std::string s1_addr = "127.0.0.1:20010";
-    std::string s2_addr = "127.0.0.1:20020";
-    std::string s3_addr = "127.0.0.1:20030";
-
-    RaftAsioPkg s1(1, s1_addr);
-    RaftAsioPkg s2(2, s2_addr);
-    RaftAsioPkg s3(3, s3_addr);
-    std::vector<RaftAsioPkg*> pkgs = {&s1, &s2, &s3};
-
-    std::atomic<size_t> write_req_cb_count(0);
-    std::atomic<size_t> write_resp_cb_count(0);
-    std::atomic<bool> start_denial(false);
-
-    _msg("launching asio-raft servers with meta callback\n");
-    for (RaftAsioPkg* rr: pkgs) {
-        rr->setMetaCallback
-            ( std::bind( test_read_meta_random_denial,
-                         &start_denial,
-                         std::placeholders::_1,
-                         std::placeholders::_2 ),
-              std::bind( test_write_req_meta,
-                         &write_req_cb_count,
-                         std::placeholders::_1 ),
-              std::bind( test_read_meta_random_denial,
-                         &start_denial,
-                         std::placeholders::_1,
-                         std::placeholders::_2 ),
-              std::bind( test_write_resp_meta,
-                         &write_resp_cb_count,
-                         std::placeholders::_1 ),
-              true );
-    }
-    CHK_Z( launch_servers(pkgs, false) );
-
-    _msg("organizing raft group\n");
-    CHK_Z( make_group(pkgs) );
-
-    TestSuite::sleep_sec(1, "wait for Raft group ready");
-
-    CHK_TRUE( s1.raftServer->is_leader() );
-    CHK_EQ(1, s1.raftServer->get_leader());
-    CHK_EQ(1, s2.raftServer->get_leader());
-    CHK_EQ(1, s3.raftServer->get_leader());
-
-    start_denial = true;
-
-    for (size_t ii=0; ii<100; ++ii) {
-        std::string msg_str = std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(sizeof(uint32_t) + msg_str.size());
-        buffer_serializer bs(msg);
-        bs.put_str(msg_str);
-        s1.raftServer->append_entries( {msg} );
-    }
-
-    TestSuite::sleep_sec(5, "wait for random denial");
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-    TestSuite::sleep_sec(1, "shutting down");
-
-    SimpleLogger::shutdown();
-    return 0;
-}
-
-
-std::string test_write_empty_meta( std::atomic<size_t>* count,
-                                   const asio_service::meta_cb_params& params )
-{
-    if (count) (*count)++;
-    return std::string();
-}
-
-bool test_read_empty_meta( std::atomic<size_t>* count,
-                           const asio_service::meta_cb_params& params,
-                           const std::string& meta )
-{
-    static std::mutex lock;
-    std::lock_guard<std::mutex> l(lock);
-
-    CHK_EQ( std::string(), meta );
-
-    if (count) (*count)++;
-    return true;
-}
-
-int empty_meta_test(bool always_invoke_cb) {
-    reset_log_files();
-
-    std::string s1_addr = "127.0.0.1:20010";
-    std::string s2_addr = "127.0.0.1:20020";
-    std::string s3_addr = "127.0.0.1:20030";
-
-    RaftAsioPkg s1(1, s1_addr);
-    RaftAsioPkg s2(2, s2_addr);
-    RaftAsioPkg s3(3, s3_addr);
-    std::vector<RaftAsioPkg*> pkgs = {&s1, &s2, &s3};
-
-    std::atomic<size_t> read_req_cb_count(0);
-    std::atomic<size_t> write_req_cb_count(0);
-    std::atomic<size_t> read_resp_cb_count(0);
-    std::atomic<size_t> write_resp_cb_count(0);
-
-    _msg("launching asio-raft servers with meta callback\n");
-    for (RaftAsioPkg* rr: pkgs) {
-        rr->setMetaCallback
-            ( std::bind( test_read_empty_meta,
-                         &read_req_cb_count,
-                         std::placeholders::_1,
-                         std::placeholders::_2 ),
-              std::bind( test_write_empty_meta,
-                         &write_req_cb_count,
-                         std::placeholders::_1 ),
-              std::bind( test_read_empty_meta,
-                         &read_resp_cb_count,
-                         std::placeholders::_1,
-                         std::placeholders::_2 ),
-              std::bind( test_write_empty_meta,
-                         &write_resp_cb_count,
-                         std::placeholders::_1 ),
-              always_invoke_cb );
-    }
-    CHK_Z( launch_servers(pkgs, false) );
-
-    _msg("organizing raft group\n");
-    CHK_Z( make_group(pkgs) );
-
-    CHK_TRUE( s1.raftServer->is_leader() );
-    CHK_EQ(1, s1.raftServer->get_leader());
-    CHK_EQ(1, s2.raftServer->get_leader());
-    CHK_EQ(1, s3.raftServer->get_leader());
-
-    TestSuite::sleep_sec(1, "wait for Raft group ready");
-
-    for (size_t ii=0; ii<10; ++ii) {
-        std::string msg_str = std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(sizeof(uint32_t) + msg_str.size());
-        buffer_serializer bs(msg);
-        bs.put_str(msg_str);
-        s1.raftServer->append_entries( {msg} );
-    }
-
-    TestSuite::sleep_sec(1, "wait for replication");
-
-    if (always_invoke_cb) {
-        // Callback functions for meta should have been called.
-        CHK_GT(read_req_cb_count.load(), 0);
-        CHK_GT(read_resp_cb_count.load(), 0);
-    } else {
-        // Callback will not be invoked on empty meta, should be 0.
-        CHK_Z(read_req_cb_count.load());
-        CHK_Z(read_resp_cb_count.load());
-    }
-    CHK_GT(write_req_cb_count, 0);
-    CHK_GT(write_resp_cb_count, 0);
-    _msg( "read req callback %zu, write req callback %zu\n",
-          read_req_cb_count.load(), write_req_cb_count.load() );
-    _msg( "read resp callback %zu, write resp callback %zu\n",
-          read_resp_cb_count.load(), write_resp_cb_count.load() );
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-    TestSuite::sleep_sec(1, "shutting down");
-
-    SimpleLogger::shutdown();
-    return 0;
-}
-
-int response_hint_test(bool with_meta) {
-    reset_log_files();
-
-    std::string s1_addr = "localhost:20010";
-    std::string s2_addr = "localhost:20020";
-    std::string s3_addr = "localhost:20030";
-
-    RaftAsioPkg s1(1, s1_addr);
-    RaftAsioPkg s2(2, s2_addr);
-    RaftAsioPkg s3(3, s3_addr);
-    std::vector<RaftAsioPkg*> pkgs = {&s1, &s2, &s3};
-
-    _msg("launching asio-raft servers %s\n",
-         with_meta ? "(with meta)" : "");
-    std::atomic<size_t> read_req_cb_count(0);
-    std::atomic<size_t> write_req_cb_count(0);
-    std::atomic<size_t> read_resp_cb_count(0);
-    std::atomic<size_t> write_resp_cb_count(0);
-    for (RaftAsioPkg* ee: pkgs) {
-        if (with_meta) {
-            ee->setMetaCallback
-                ( std::bind( test_read_req_meta,
-                             &read_req_cb_count,
-                             std::placeholders::_1,
-                             std::placeholders::_2 ),
-                  std::bind( test_write_req_meta,
-                             &write_req_cb_count,
-                             std::placeholders::_1 ),
-                  std::bind( test_read_resp_meta,
-                             &read_resp_cb_count,
-                             std::placeholders::_1,
-                             std::placeholders::_2 ),
-                  std::bind( test_write_resp_meta,
-                             &write_resp_cb_count,
-                             std::placeholders::_1 ),
-                  true );
-        }
-    }
-    CHK_Z( launch_servers(pkgs, false) );
-
-    _msg("enable batch size hint with positive value\n");
-    for (RaftAsioPkg* ee: pkgs) {
-        ee->getTestSm()->set_next_batch_size_hint_in_bytes(1);
-    }
-
-    _msg("organizing raft group\n");
-    CHK_Z( make_group(pkgs) );
-
-    CHK_TRUE( s1.raftServer->is_leader() );
-    CHK_EQ(1, s1.raftServer->get_leader());
-    CHK_EQ(1, s2.raftServer->get_leader());
-    CHK_EQ(1, s3.raftServer->get_leader());
-    TestSuite::sleep_sec(1, "wait for Raft group ready");
-
-    const size_t NUM = 100;
-    for (size_t ii=0; ii<NUM; ++ii) {
-        std::string msg_str = std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(sizeof(uint32_t) + msg_str.size());
-        buffer_serializer bs(msg);
-        bs.put_str(msg_str);
-        s1.raftServer->append_entries( {msg} );
-    }
-    TestSuite::sleep_sec(1, "wait for replication");
-
-    // State machine should be identical.
-    CHK_OK( s2.getTestSm()->isSame( *s1.getTestSm() ) );
-    CHK_OK( s3.getTestSm()->isSame( *s1.getTestSm() ) );
-
-    _msg("disable batch size hint\n");
-    for (RaftAsioPkg* ee: pkgs) {
-        ee->getTestSm()->set_next_batch_size_hint_in_bytes(0);
-    }
-
-    for (size_t ii=0; ii<NUM; ++ii) {
-        std::string msg_str = "2nd_" + std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(sizeof(uint32_t) + msg_str.size());
-        buffer_serializer bs(msg);
-        bs.put_str(msg_str);
-        s1.raftServer->append_entries( {msg} );
-    }
-    TestSuite::sleep_sec(1, "wait for replication");
-
-    // State machine should be identical.
-    CHK_OK( s2.getTestSm()->isSame( *s1.getTestSm() ) );
-    CHK_OK( s3.getTestSm()->isSame( *s1.getTestSm() ) );
-
-    _msg("enable batch size hint with negative value\n");
-    for (RaftAsioPkg* ee: pkgs) {
-        ee->getTestSm()->set_next_batch_size_hint_in_bytes(-1);
-    }
-
-    TestSuite::sleep_sec(1, "wait peer's hint size info refreshed in leader side");
-
-    // With negative hint size, append_entries will timeout due to
-    // raft server can not commit. Set timeout to a small value.
-    raft_params params = s1.raftServer->get_current_params();
-    params.with_client_req_timeout(1000);
-    s1.raftServer->update_params(params);
-
-    for (size_t ii=0; ii<3; ++ii) {
-        std::string msg_str = "3rd_" + std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(sizeof(uint32_t) + msg_str.size());
-        buffer_serializer bs(msg);
-        bs.put_str(msg_str);
-        s1.raftServer->append_entries( {msg} );
-    }
-    TestSuite::sleep_sec(1, "wait for replication but actually no replication happen");
-
-    // State machine should be identical. All are not committed.
-    CHK_OK( s2.getTestSm()->isSame( *s1.getTestSm() ) );
-    CHK_OK( s3.getTestSm()->isSame( *s1.getTestSm() ) );
-
-    if (with_meta) {
-        // Callback functions for meta should have been called.
-        CHK_GT(read_req_cb_count.load(), 0);
-        CHK_GT(write_req_cb_count.load(), 0);
-        CHK_GT(read_resp_cb_count.load(), 0);
-        CHK_GT(write_resp_cb_count.load(), 0);
-        _msg( "read req callback %zu, write req callback %zu\n",
-              read_req_cb_count.load(), write_req_cb_count.load() );
-        _msg( "read resp callback %zu, write resp callback %zu\n",
-              read_resp_cb_count.load(), write_resp_cb_count.load() );
-    }
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-    TestSuite::sleep_sec(1, "shutting down");
-
-    SimpleLogger::shutdown();
-    return 0;
-}
-
-static void async_handler(std::list<ulong>* idx_list,
-                          std::mutex* idx_list_lock,
-                          ptr<buffer>& result,
-                          ptr<std::exception>& err)
-{
-    result->pos(0);
-    ulong idx = result->get_ulong();
-    if (idx_list) {
-        std::lock_guard<std::mutex> l(*idx_list_lock);
-        idx_list->push_back(idx);
-    }
-}
-
 int async_append_handler_test() {
     reset_log_files();
 
@@ -1021,216 +425,6 @@ int async_append_handler_with_order_inversion_test() {
     s1.raftServer->shutdown();
     s2.raftServer->shutdown();
     s3.raftServer->shutdown();
-    TestSuite::sleep_sec(1, "shutting down");
-
-    SimpleLogger::shutdown();
-    return 0;
-}
-
-int auto_quorum_size_test() {
-    reset_log_files();
-
-    std::string s1_addr = "tcp://127.0.0.1:20010";
-    std::string s2_addr = "tcp://127.0.0.1:20020";
-
-    RaftAsioPkg s1(1, s1_addr);
-    std::shared_ptr<RaftAsioPkg> s2 = std::make_shared<RaftAsioPkg>(2, s2_addr);
-    std::vector<RaftAsioPkg*> pkgs = {&s1, s2.get()};
-
-    _msg("launching asio-raft servers\n");
-    CHK_Z( launch_servers(pkgs, false) );
-
-    _msg("organizing raft group\n");
-    CHK_Z( make_group(pkgs) );
-    TestSuite::sleep_sec(1, "wait for Raft group ready");
-
-    // Set custom term counter, and enable auto quorum size mode.
-    auto custom_inc_term = [](uint64_t cur_term) -> uint64_t {
-        return (cur_term / 10) + 10;
-    };
-    s1.raftServer->set_inc_term_func(custom_inc_term);
-    s2->raftServer->set_inc_term_func(custom_inc_term);
-
-    raft_params params = s1.raftServer->get_current_params();
-    params.auto_adjust_quorum_for_small_cluster_ = true;
-    s1.raftServer->update_params(params);
-    s2->raftServer->update_params(params);
-
-    CHK_TRUE( s1.raftServer->is_leader() );
-    CHK_EQ(1, s1.raftServer->get_leader());
-    CHK_EQ(1, s2->raftServer->get_leader());
-
-    // Replication.
-    for (size_t ii=0; ii<10; ++ii) {
-        std::string msg_str = std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(sizeof(uint32_t) + msg_str.size());
-        buffer_serializer bs(msg);
-        bs.put_str(msg_str);
-        s1.raftServer->append_entries( {msg} );
-    }
-    TestSuite::sleep_sec(1, "wait for replication");
-    uint64_t committed_idx = s1.raftServer->get_committed_log_idx();
-
-    // State machine should be identical.
-    CHK_OK( s2->getTestSm()->isSame( *s1.getTestSm() ) );
-
-    // Shutdown S2.
-    s2->raftServer->shutdown();
-    s2.reset();
-
-    TestSuite::sleep_ms( RaftAsioPkg::HEARTBEAT_MS * 30,
-                         "wait for quorum adjust" );
-
-    // More replication.
-    for (size_t ii=10; ii<11; ++ii) {
-        std::string msg_str = std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(sizeof(uint32_t) + msg_str.size());
-        buffer_serializer bs(msg);
-        bs.put_str(msg_str);
-        s1.raftServer->append_entries( {msg} );
-    }
-
-    // Replication should succeed: committed index should be moved forward.
-    TestSuite::sleep_sec(1, "wait for replication");
-    CHK_EQ( committed_idx + 1,
-            s1.raftServer->get_committed_log_idx() );
-
-    // Restart S2.
-    _msg("launching S2 again\n");
-    RaftAsioPkg s2_new(2, s2_addr);
-    CHK_Z( launch_servers({&s2_new}, false) );
-    TestSuite::sleep_sec(1, "wait for S2 ready");
-    CHK_EQ( committed_idx + 1,
-            s2_new.raftServer->get_committed_log_idx() );
-
-    // More replication.
-    for (size_t ii=11; ii<12; ++ii) {
-        std::string msg_str = std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(sizeof(uint32_t) + msg_str.size());
-        buffer_serializer bs(msg);
-        bs.put_str(msg_str);
-        s1.raftServer->append_entries( {msg} );
-    }
-
-    // Both of them should have the same commit number.
-    TestSuite::sleep_sec(1, "wait for replication");
-    CHK_EQ( committed_idx + 2,
-            s1.raftServer->get_committed_log_idx() );
-    CHK_EQ( committed_idx + 2,
-            s2_new.raftServer->get_committed_log_idx() );
-
-    s1.raftServer->shutdown();
-    s2_new.raftServer->shutdown();
-    TestSuite::sleep_sec(1, "shutting down");
-
-    SimpleLogger::shutdown();
-    return 0;
-}
-
-int auto_quorum_size_election_test() {
-    reset_log_files();
-
-    std::string s1_addr = "tcp://127.0.0.1:20010";
-    std::string s2_addr = "tcp://127.0.0.1:20020";
-
-    std::shared_ptr<RaftAsioPkg> s1 = std::make_shared<RaftAsioPkg>(1, s1_addr);
-    std::shared_ptr<RaftAsioPkg> s2 = std::make_shared<RaftAsioPkg>(2, s2_addr);
-    std::vector<RaftAsioPkg*> pkgs = {s1.get(), s2.get()};
-
-    _msg("launching asio-raft servers\n");
-    CHK_Z( launch_servers(pkgs, false) );
-
-    _msg("organizing raft group\n");
-    CHK_Z( make_group(pkgs) );
-    TestSuite::sleep_sec(1, "wait for Raft group ready");
-
-    // Set custom term counter, and enable auto quorum size mode.
-    auto custom_inc_term = [](uint64_t cur_term) -> uint64_t {
-        return (cur_term / 10) + 10;
-    };
-    s1->raftServer->set_inc_term_func(custom_inc_term);
-    s2->raftServer->set_inc_term_func(custom_inc_term);
-
-    raft_params params = s1->raftServer->get_current_params();
-    params.auto_adjust_quorum_for_small_cluster_ = true;
-    s1->raftServer->update_params(params);
-    s2->raftServer->update_params(params);
-
-    CHK_TRUE( s1->raftServer->is_leader() );
-    CHK_EQ(1, s1->raftServer->get_leader());
-    CHK_EQ(1, s2->raftServer->get_leader());
-
-    // Replication.
-    for (size_t ii=0; ii<10; ++ii) {
-        std::string msg_str = std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(sizeof(uint32_t) + msg_str.size());
-        buffer_serializer bs(msg);
-        bs.put_str(msg_str);
-        s1->raftServer->append_entries( {msg} );
-    }
-    TestSuite::sleep_sec(1, "wait for replication");
-
-    // State machine should be identical.
-    CHK_OK( s2->getTestSm()->isSame( *s1->getTestSm() ) );
-
-    // Shutdown S1.
-    s1->raftServer->shutdown();
-    s1.reset();
-
-    // Wait for adjust quorum and self election.
-    TestSuite::sleep_ms( RaftAsioPkg::HEARTBEAT_MS * 50,
-                         "wait for quorum adjust" );
-
-    // S2 should be a leader.
-    CHK_TRUE( s2->raftServer->is_leader() );
-    CHK_EQ(2, s2->raftServer->get_leader());
-    uint64_t committed_idx = s2->raftServer->get_committed_log_idx();
-
-    // More replication.
-    for (size_t ii=10; ii<11; ++ii) {
-        std::string msg_str = std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(sizeof(uint32_t) + msg_str.size());
-        buffer_serializer bs(msg);
-        bs.put_str(msg_str);
-        s2->raftServer->append_entries( {msg} );
-    }
-
-    // Replication should succeed: committed index should be moved forward.
-    TestSuite::sleep_sec(1, "wait for replication");
-    CHK_EQ( committed_idx + 1,
-            s2->raftServer->get_committed_log_idx() );
-
-    // Restart S1.
-    _msg("launching S1 again\n");
-    RaftAsioPkg s1_new(1, s1_addr);
-    CHK_Z( launch_servers({&s1_new}, false) );
-    TestSuite::sleep_sec(1, "wait for S2 ready");
-    CHK_EQ( committed_idx + 1,
-            s1_new.raftServer->get_committed_log_idx() );
-
-    // S2 should remain as a leader.
-    CHK_TRUE( s2->raftServer->is_leader() );
-    CHK_EQ(2, s1_new.raftServer->get_leader());
-    CHK_EQ(2, s2->raftServer->get_leader());
-
-    // More replication.
-    for (size_t ii=11; ii<12; ++ii) {
-        std::string msg_str = std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(sizeof(uint32_t) + msg_str.size());
-        buffer_serializer bs(msg);
-        bs.put_str(msg_str);
-        s2->raftServer->append_entries( {msg} );
-    }
-
-    // Both of them should have the same commit number.
-    TestSuite::sleep_sec(1, "wait for replication");
-    CHK_EQ( committed_idx + 2,
-            s1_new.raftServer->get_committed_log_idx() );
-    CHK_EQ( committed_idx + 2,
-            s2->raftServer->get_committed_log_idx() );
-
-    s2->raftServer->shutdown();
-    s1_new.raftServer->shutdown();
     TestSuite::sleep_sec(1, "shutting down");
 
     SimpleLogger::shutdown();
@@ -2330,95 +1524,6 @@ int pause_state_machine_execution_test(bool use_global_mgr) {
     return 0;
 }
 
-int full_consensus_test() {
-    reset_log_files();
-
-    std::string s1_addr = "tcp://127.0.0.1:20010";
-    std::string s2_addr = "tcp://127.0.0.1:20020";
-    std::string s3_addr = "tcp://127.0.0.1:20030";
-
-    RaftAsioPkg s1(1, s1_addr);
-    RaftAsioPkg s2(2, s2_addr);
-    RaftAsioPkg s3(3, s3_addr);
-    std::vector<RaftAsioPkg*> pkgs = {&s1, &s2, &s3};
-
-    _msg("launching asio-raft servers\n");
-    CHK_Z( launch_servers(pkgs, false) );
-
-    _msg("organizing raft group\n");
-    CHK_Z( make_group(pkgs) );
-
-    // Set async & full consensus mode.
-    for (auto& entry: pkgs) {
-        RaftAsioPkg* pp = entry;
-        raft_params param = pp->raftServer->get_current_params();
-        param.return_method_ = raft_params::async_handler;
-        param.use_full_consensus_among_healthy_members_ = true;
-        pp->raftServer->update_params(param);
-    }
-
-    // Stop S3.
-    s3.raftServer->shutdown();
-    s3.stopAsio();
-    TestSuite::sleep_ms(RaftAsioPkg::HEARTBEAT_MS * 5, "stop S3");
-
-    // Remember the commit index.
-    uint64_t commit_idx = s1.raftServer->get_committed_log_idx();
-
-    // Append messages asynchronously.
-    const size_t NUM = 10;
-    std::list< ptr< cmd_result< ptr<buffer> > > > handlers;
-    std::list<ulong> idx_list;
-    std::mutex idx_list_lock;
-    auto do_async_append = [&]() {
-        handlers.clear();
-        idx_list.clear();
-        for (size_t ii=0; ii<NUM; ++ii) {
-            std::string test_msg = "test" + std::to_string(ii);
-            ptr<buffer> msg = buffer::alloc(test_msg.size() + 1);
-            msg->put(test_msg);
-            ptr< cmd_result< ptr<buffer> > > ret =
-                s1.raftServer->append_entries( {msg} );
-
-            cmd_result< ptr<buffer> >::handler_type my_handler =
-                std::bind( async_handler,
-                           &idx_list,
-                           &idx_list_lock,
-                           std::placeholders::_1,
-                           std::placeholders::_2 );
-            ret->when_ready( my_handler );
-
-            handlers.push_back(ret);
-        }
-    };
-    do_async_append();
-
-    TestSuite::sleep_ms(RaftAsioPkg::HEARTBEAT_MS * 5, "wait for replication");
-
-    // No request should have been committed.
-    CHK_EQ(commit_idx, s1.raftServer->get_committed_log_idx());
-
-    // Wait more so that the leader can tolerate not responding peer.
-    TestSuite::sleep_ms(RaftAsioPkg::HEARTBEAT_MS * 15, "wait for not responding peer");
-    uint64_t new_commit_idx = s1.raftServer->get_committed_log_idx();
-    CHK_GT(new_commit_idx, commit_idx);
-
-    // More replication.
-    do_async_append();
-    TestSuite::sleep_ms(RaftAsioPkg::HEARTBEAT_MS * 5, "wait for replication");
-    // They should be committed immediately.
-    CHK_GT(s1.raftServer->get_committed_log_idx(), new_commit_idx);
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-    TestSuite::sleep_sec(1, "shutting down");
-
-    SimpleLogger::shutdown();
-    return 0;
-}
-
-
 int flip_learner_flag_test() {
     reset_log_files();
 
@@ -2460,96 +1565,6 @@ int flip_learner_flag_test() {
         ptr<cluster_config> conf = pp->raftServer->get_config();
         CHK_FALSE(conf->get_server(s3.myId)->is_learner());
     }
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-    TestSuite::sleep_sec(1, "shutting down");
-
-    SimpleLogger::shutdown();
-    return 0;
-}
-
-int custom_commit_condition_test() {
-    reset_log_files();
-
-    std::string s1_addr = "tcp://127.0.0.1:20010";
-    std::string s2_addr = "tcp://127.0.0.1:20020";
-    std::string s3_addr = "tcp://127.0.0.1:20030";
-
-    RaftAsioPkg s1(1, s1_addr);
-    RaftAsioPkg s2(2, s2_addr);
-    RaftAsioPkg s3(3, s3_addr);
-    std::vector<RaftAsioPkg*> pkgs = {&s1, &s2, &s3};
-
-    _msg("launching asio-raft servers\n");
-    CHK_Z( launch_servers(pkgs, false) );
-
-    _msg("organizing raft group\n");
-    CHK_Z( make_group(pkgs) );
-
-    // Set async & full consensus mode.
-    for (auto& entry: pkgs) {
-        RaftAsioPkg* pp = entry;
-        raft_params param = pp->raftServer->get_current_params();
-        param.return_method_ = raft_params::async_handler;
-        pp->raftServer->update_params(param);
-    }
-
-    // Stop S3.
-    s3.raftServer->shutdown();
-    s3.stopAsio();
-    TestSuite::sleep_ms(RaftAsioPkg::HEARTBEAT_MS * 5, "stop S3");
-
-    // Remember the commit index.
-    uint64_t commit_idx = s1.raftServer->get_committed_log_idx();
-
-    // Set custom quorum set: {S1, S3}.
-    s1.getTestSm()->setServersForCommit({1, 3});
-
-    // Append messages asynchronously.
-    const size_t NUM = 10;
-    std::list< ptr< cmd_result< ptr<buffer> > > > handlers;
-    std::list<ulong> idx_list;
-    std::mutex idx_list_lock;
-    auto do_async_append = [&]() {
-        handlers.clear();
-        idx_list.clear();
-        for (size_t ii=0; ii<NUM; ++ii) {
-            std::string test_msg = "test" + std::to_string(ii);
-            ptr<buffer> msg = buffer::alloc(test_msg.size() + 1);
-            msg->put(test_msg);
-            ptr< cmd_result< ptr<buffer> > > ret =
-                s1.raftServer->append_entries( {msg} );
-
-            cmd_result< ptr<buffer> >::handler_type my_handler =
-                std::bind( async_handler,
-                           &idx_list,
-                           &idx_list_lock,
-                           std::placeholders::_1,
-                           std::placeholders::_2 );
-            ret->when_ready( my_handler );
-
-            handlers.push_back(ret);
-        }
-    };
-    do_async_append();
-
-    TestSuite::sleep_sec(1, "wait for replication");
-
-    // No request should have been committed, as S1 cannot reach quorum due to S3.
-    CHK_EQ(commit_idx, s1.raftServer->get_committed_log_idx());
-
-    // Restart S3.
-    raft_params new_params = s1.raftServer->get_current_params();
-    s3.restartServer(&new_params);
-    TestSuite::sleep_ms(500, "restarting S3");
-
-    // More replication.
-    do_async_append();
-    TestSuite::sleep_sec(1, "wait for replication");
-    // They should be committed immediately.
-    CHK_GT(s1.raftServer->get_committed_log_idx(), commit_idx);
 
     s1.raftServer->shutdown();
     s2.raftServer->shutdown();
@@ -3001,32 +2016,11 @@ int main(int argc, char** argv) {
                ssl_test );
 #endif
 
-    ts.doTest( "message meta test",
-               message_meta_test,
-               TestRange<bool>( {false, true} ) );
-
-    ts.doTest( "empty meta test",
-               empty_meta_test,
-               TestRange<bool>( {false, true} ) );
-
-    ts.doTest( "message meta random denial test",
-               message_meta_random_denial_test );
-
-    ts.doTest( "response hint test",
-               response_hint_test,
-               TestRange<bool>( {false, true} ) );
-
     ts.doTest( "async append handler test",
                async_append_handler_test );
 
     ts.doTest( "async append handler with order inversion test",
                async_append_handler_with_order_inversion_test );
-
-    ts.doTest( "auto quorum size test",
-               auto_quorum_size_test );
-
-    ts.doTest( "auto quorum size for election test",
-               auto_quorum_size_election_test );
 
     ts.doTest( "global manager basic test",
                global_mgr_basic_test );
@@ -3076,14 +2070,8 @@ int main(int argc, char** argv) {
                pause_state_machine_execution_test,
                TestRange<bool>( {false, true} ) );
 
-    ts.doTest( "full consensus test",
-               full_consensus_test );
-
     ts.doTest( "flip learner flag test",
                flip_learner_flag_test);
-
-    ts.doTest( "custom commit condition test",
-               custom_commit_condition_test );
 
     ts.doTest( "parallel log append test",
                parallel_log_append_test );
