+++ tests/unit/raft_server_test.cxx
@@ -1,5 +1,5 @@
 /************************************************************************
-Copyright 2017-2019 eBay Inc.
+Copyright 2017-present eBay Inc.
 Author/Developer(s): Jung-Sang Ahn
 
 Licensed under the Apache License, Version 2.0 (the "License");
@@ -18,6 +18,7 @@ limitations under the License.
 #include "debugging_options.hxx"
 #include "fake_network.hxx"
 #include "raft_package_fake.hxx"
+#include "fake_executer.hxx"
 
 #include "event_awaiter.hxx"
 #include "raft_params.hxx"
@@ -30,73 +31,7 @@ using namespace raft_functional_common;
 
 using raft_result = cmd_result< ptr<buffer> >;
 
-std::ostream& operator<<(std::ostream& os, raft_server::PrioritySetResult res) {
-    if (res == raft_server::PrioritySetResult::SET)
-        os << "SET";
-    else if (res == raft_server::PrioritySetResult::IGNORED)
-        os << "IGNORED";
-    else
-        os << "BROADCAST";
-    return os;
-}
-
 namespace raft_server_test {
-
-struct ExecArgs : TestSuite::ThreadArgs {
-    ExecArgs(RaftPkg* _leader)
-        : leader(_leader)
-        , stopSignal(false)
-        , msgToWrite(nullptr)
-        {}
-
-    void setMsg(ptr<buffer>& to) {
-        std::lock_guard<std::mutex> l(msgToWriteLock);
-        msgToWrite = to;
-    }
-
-    ptr<buffer> getMsg() {
-        std::lock_guard<std::mutex> l(msgToWriteLock);
-        return msgToWrite;
-    }
-
-    RaftPkg* leader;
-    std::atomic<bool> stopSignal;
-    ptr<buffer> msgToWrite;
-    std::mutex msgToWriteLock;
-    EventAwaiter eaExecuter;
-};
-
-static size_t EXECUTOR_WAIT_MS = 100;
-
-// Mimic the user of Raft server, which has a separate executer thread.
-int fake_executer(TestSuite::ThreadArgs* _args) {
-    ExecArgs* args = static_cast<ExecArgs*>(_args);
-
-    while (!args->stopSignal) {
-        args->eaExecuter.wait_ms(10000);
-        args->eaExecuter.reset();
-        if (args->stopSignal) break;
-
-        ptr<buffer> msg = nullptr;
-        {   std::lock_guard<std::mutex> l(args->msgToWriteLock);
-            if (!args->msgToWrite) continue;
-            msg = args->msgToWrite;
-            args->msgToWrite.reset();
-        }
-
-        args->leader->dbgLog(" --- append ---");
-        args->leader->raftServer->append_entries( {msg} );
-    }
-
-    return 0;
-}
-
-int fake_executer_killer(TestSuite::ThreadArgs* _args) {
-    ExecArgs* args = static_cast<ExecArgs*>(_args);
-    args->stopSignal = true;
-    args->eaExecuter.invoke();
-    return 0;
-}
 
 int make_group_test() {
     reset_log_files();
@@ -857,885 +792,6 @@ int multiple_config_change_test() {
     return 0;
 }
 
-int leader_election_basic_test() {
-    reset_log_files();
-    ptr<FakeNetworkBase> f_base = cs_new<FakeNetworkBase>();
-
-    std::string s1_addr = "S1";
-    std::string s2_addr = "S2";
-    std::string s3_addr = "S3";
-
-    RaftPkg s1(f_base, 1, s1_addr);
-    RaftPkg s2(f_base, 2, s2_addr);
-    RaftPkg s3(f_base, 3, s3_addr);
-    std::vector<RaftPkg*> pkgs = {&s1, &s2, &s3};
-
-    CHK_Z( launch_servers( pkgs ) );
-    CHK_Z( make_group( pkgs ) );
-
-    // Keep the last log index.
-    uint64_t last_idx = s1.raftServer->get_last_log_idx();
-
-    // Trigger election timer of S2.
-    s2.dbgLog(" --- invoke election timer of S2 ---");
-    s2.fTimer->invoke( timer_task_type::election_timer );
-    // Send pre-vote requests, and probably rejected by S1 and S3.
-    s2.fNet->execReqResp();
-
-    // Trigger election timer of S3.
-    s3.dbgLog(" --- invoke election timer of S3 ---");
-    s3.fTimer->invoke( timer_task_type::election_timer );
-
-    // Send pre-vote requests, it will be rejected by S1, accepted by S2.
-    // As a part of resp handling, it will initiate vote request.
-    s3.fNet->execReqResp();
-    // Send vote requests, S3 will be elected as a leader.
-    s3.fNet->execReqResp();
-    // Wait for bg commit for configuration change.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Send new config as a new leader.
-    s3.fNet->execReqResp();
-    // Follow-up: commit.
-    s3.fNet->execReqResp();
-    // Wait for bg commit for configuration change.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // S3's log index at becoming leader should be the next index of
-    // the last one.
-    TestSuite::Msg mm;
-    uint64_t idx_at_leader = s3.raftServer->get_log_idx_at_becoming_leader();
-    mm << "last index " << last_idx << ", log index at becoming leader "
-       << idx_at_leader << std::endl;
-    CHK_EQ( last_idx + 1, idx_at_leader );
-
-    // S1 and S2's log index should be 0.
-    CHK_Z( s1.raftServer->get_log_idx_at_becoming_leader() );
-    CHK_Z( s2.raftServer->get_log_idx_at_becoming_leader() );
-
-    CHK_FALSE( s1.raftServer->is_leader() );
-    CHK_FALSE( s2.raftServer->is_leader() );
-    CHK_TRUE( s3.raftServer->is_leader() );
-
-    print_stats(pkgs);
-
-    s3.raftServer->shutdown();
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-
-    f_base->destroy();
-
-    return 0;
-}
-
-int leader_election_priority_test() {
-    reset_log_files();
-    ptr<FakeNetworkBase> f_base = cs_new<FakeNetworkBase>();
-
-    std::string s1_addr = "S1";
-    std::string s2_addr = "S2";
-    std::string s3_addr = "S3";
-
-    RaftPkg s1(f_base, 1, s1_addr);
-    RaftPkg s2(f_base, 2, s2_addr);
-    RaftPkg s3(f_base, 3, s3_addr);
-    std::vector<RaftPkg*> pkgs = {&s1, &s2, &s3};
-
-    CHK_Z( launch_servers( pkgs ) );
-    CHK_Z( make_group( pkgs ) );
-
-    // Set the priority of S2 to 100.
-    CHK_EQ( raft_server::PrioritySetResult::SET, s1.raftServer->set_priority(2, 100) );
-    // Send priority change reqs.
-    s1.fNet->execReqResp();
-    // Send reqs again for commit.
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Set the priority of S3 to 85.
-    CHK_EQ( raft_server::PrioritySetResult::SET, s1.raftServer->set_priority(3, 85) );
-    // Send priority change reqs.
-    s1.fNet->execReqResp();
-    // Send reqs again for commit.
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Trigger election timer of S2.
-    s2.dbgLog(" --- invoke election timer of S2 ---");
-    s2.fTimer->invoke( timer_task_type::election_timer );
-    // Send pre-vote requests, and probably rejected by S1 and S3.
-    s2.fNet->execReqResp();
-
-    // Trigger election timer of S3.
-    s3.dbgLog(" --- invoke election timer of S3 ---");
-    // It will not initiate vote due to priority.
-    s3.fTimer->invoke( timer_task_type::election_timer );
-
-    // Trigger election timer of S3 again.
-    s3.dbgLog(" --- invoke election timer of S3 ---");
-    // Now it will initiate vote by help of priority decay.
-    s3.fTimer->invoke( timer_task_type::election_timer );
-
-    // Send pre-vote requests, it will be rejected by S1, accepted by S2.
-    // As a part of resp handling, it will initiate vote request.
-    s3.fNet->execReqResp();
-    // Send vote requests, S2 will deny due to priority.
-    s3.fNet->execReqResp();
-
-    // Trigger election timer of S2.
-    s2.dbgLog(" --- invoke election timer of S2 ---");
-    s2.fTimer->invoke( timer_task_type::election_timer );
-    // Send pre-vote requests, and accepted by S3.
-    // As a result of response, it will initiate actual vote.
-    s2.fNet->execReqResp();
-    // Send vote requests, S3 will vote for it.
-    s2.fNet->execReqResp();
-    // Wait for bg commit for configuration change.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Send new config as a new leader.
-    s2.fNet->execReqResp();
-    // Follow-up: commit.
-    s2.fNet->execReqResp();
-    // Wait for bg commit for configuration change.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    CHK_FALSE( s1.raftServer->is_leader() );
-    CHK_TRUE( s2.raftServer->is_leader() );
-    CHK_FALSE( s3.raftServer->is_leader() );
-
-    print_stats(pkgs);
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-
-    f_base->destroy();
-
-    return 0;
-}
-
-int leader_election_with_aggressive_node_test() {
-    reset_log_files();
-    ptr<FakeNetworkBase> f_base = cs_new<FakeNetworkBase>();
-
-    std::string s1_addr = "S1";
-    std::string s2_addr = "S2";
-    std::string s3_addr = "S3";
-
-    RaftPkg s1(f_base, 1, s1_addr);
-    RaftPkg s2(f_base, 2, s2_addr);
-    RaftPkg s3(f_base, 3, s3_addr);
-    std::vector<RaftPkg*> pkgs = {&s1, &s2, &s3};
-
-    CHK_Z( launch_servers( pkgs ) );
-    CHK_Z( make_group( pkgs ) );
-
-    // Set the priority of S1 to 100.
-    CHK_EQ( raft_server::PrioritySetResult::SET, s1.raftServer->set_priority(1, 100) );
-    s1.fNet->execReqResp();
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Set the priority of S2 to 50.
-    CHK_EQ( raft_server::PrioritySetResult::SET, s1.raftServer->set_priority(2, 50) );
-    s1.fNet->execReqResp();
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Set the priority of S3 to 1.
-    CHK_EQ( raft_server::PrioritySetResult::SET, s1.raftServer->set_priority(3, 1) );
-    s1.fNet->execReqResp();
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // --- Now assume S1 is not reachable. ---
-
-    // Trigger election timer of S2.
-    s2.dbgLog(" --- invoke election timer of S2 ---");
-    // It will not initiate vote due to priority.
-    s2.fTimer->invoke( timer_task_type::election_timer );
-
-    // Keep triggering election timer of S3, until it becomes leader.
-    const size_t MAX_ATTEMPTS = 1000;
-    size_t num_attempts = 0;
-    TestSuite::UnknownProgress pp("leader election attempts: ");
-    do {
-        s3.dbgLog(" --- invoke election timer of S3 ---");
-        s3.fTimer->invoke( timer_task_type::election_timer );
-        // Drop all packets to S1.
-        s3.fNet->makeReqFailAll(s1_addr);
-        // Send pre-vote requests.
-        s3.fNet->execReqResp();
-        // Send vote request (if exists).
-        s3.fNet->execReqResp();
-        pp.update(++num_attempts);
-    } while ( !s3.raftServer->is_leader() &&
-              num_attempts < MAX_ATTEMPTS );
-    pp.done();
-    _msg("%zu attempts\n", num_attempts);
-    CHK_SM(num_attempts, MAX_ATTEMPTS);
-
-    // Wait for bg commit for configuration change.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    s3.fNet->makeReqFailAll(s1_addr);
-    s3.fNet->execReqResp();
-    s3.fNet->execReqResp();
-    // Wait for bg commit for configuration change.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    CHK_FALSE( s2.raftServer->is_leader() );
-    CHK_TRUE( s3.raftServer->is_leader() );
-
-    print_stats(pkgs);
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-
-    f_base->destroy();
-
-    return 0;
-}
-int leader_election_with_catching_up_server_test() {
-    reset_log_files();
-    ptr<FakeNetworkBase> f_base = cs_new<FakeNetworkBase>();
-
-    std::string s1_addr = "S1";
-    std::string s2_addr = "S2";
-
-    RaftPkg s1(f_base, 1, s1_addr);
-    RaftPkg s2(f_base, 2, s2_addr);
-    std::vector<RaftPkg*> pkgs = {&s1, &s2};
-
-    CHK_Z( launch_servers( pkgs ) );
-
-    // Set priority of S1 to 80 and S2 to 100.
-    // Each server thinks it's a leader
-    CHK_EQ( raft_server::PrioritySetResult::SET, s1.raftServer->set_priority(1, 80) );
-    CHK_EQ( raft_server::PrioritySetResult::SET, s2.raftServer->set_priority(2, 100) );
-
-    // Append logs to S1 to trigger log compaction.
-    for (auto& entry: pkgs) {
-        RaftPkg* pp = entry;
-        raft_params param = pp->raftServer->get_current_params();
-        param.return_method_ = raft_params::async_handler;
-        param.log_sync_stop_gap_ = 0;
-        pp->raftServer->update_params(param);
-    }
-    const size_t NUM = 10;
-    std::list< ptr< cmd_result< ptr<buffer> > > > handlers;
-    for (size_t ii=0; ii<NUM; ++ii) {
-        std::string test_msg = "test" + std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(test_msg.size() + 1);
-        msg->put(test_msg);
-        ptr< cmd_result< ptr<buffer> > > ret =
-            s1.raftServer->append_entries( {msg} );
-
-        CHK_TRUE( ret->get_accepted() );
-
-        handlers.push_back(ret);
-    }
-    // Pre-commit and commit.
-    s1.fNet->execReqResp();
-    s1.fNet->execReqResp();
-    // Wait for bg commit.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // All handlers should be OK.
-    for (auto& entry: handlers) {
-        CHK_TRUE( entry->has_result() );
-        CHK_EQ( cmd_result_code::OK, entry->get_result_code() );
-    }
-
-    // Add S2 to S1.
-    s1.raftServer->add_srv( *s2.getTestMgr()->get_srv_config() );
-
-    // Join req/resp.
-    s1.fNet->execReqResp();
-    // Add new server, notify existing peers.
-    // After getting response, it will make configuration commit.
-    s1.fNet->execReqResp();
-    // Notify new commit.
-    s1.fNet->execReqResp();
-    // Wait for bg commit for configuration change.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Now S2 is the member of the cluster.
-
-    // S1 resigns immediately (to schedule election timer).
-    s1.raftServer->yield_leadership(true);
-    CHK_FALSE( s1.raftServer->is_leader() );
-
-    // Invoke election timer of S1.
-    s1.dbgLog(" --- invoke election timer of S1 ---");
-    s1.fTimer->invoke( timer_task_type::election_timer );
-    // Send pre-vote request, S2 should accept it as it is in catch-up mode.
-    s1.fNet->execReqResp();
-    // Send vote request, S2 should accept it as it is in catch-up mode.
-    s1.fNet->execReqResp();
-
-    // Leader election should succeed.
-    CHK_TRUE( s1.raftServer->is_leader() );
-
-    print_stats(pkgs);
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-
-    f_base->destroy();
-
-    return 0;
-}
-
-int leadership_takeover_basic_test() {
-    reset_log_files();
-    ptr<FakeNetworkBase> f_base = cs_new<FakeNetworkBase>();
-
-    std::string s1_addr = "S1";
-    std::string s2_addr = "S2";
-    std::string s3_addr = "S3";
-
-    RaftPkg s1(f_base, 1, s1_addr);
-    RaftPkg s2(f_base, 2, s2_addr);
-    RaftPkg s3(f_base, 3, s3_addr);
-    std::vector<RaftPkg*> pkgs = {&s1, &s2, &s3};
-
-    CHK_Z( launch_servers( pkgs ) );
-    CHK_Z( make_group( pkgs ) );
-
-    // Set the priority of S2 to 10.
-    CHK_EQ( raft_server::PrioritySetResult::SET, s1.raftServer->set_priority(2, 10) );
-    // Send priority change reqs.
-    s1.fNet->execReqResp();
-    // Send reqs again for commit.
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Set the priority of S3 to 5.
-    CHK_EQ( raft_server::PrioritySetResult::SET, s1.raftServer->set_priority(3, 5) );
-    // Send priority change reqs.
-    s1.fNet->execReqResp();
-    // Send reqs again for commit.
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Yield leadership.
-    s1.dbgLog(" --- yield leadership ---");
-    s1.raftServer->yield_leadership();
-    // Send heartbeat.
-    s1.fTimer->invoke( timer_task_type::heartbeat_timer );
-    s1.fNet->execReqResp();
-    // After getting response of heartbeat, S1 will resign.
-    s1.fNet->execReqResp();
-
-    // Now S2 should have received takeover request.
-    // Send vote requests.
-    s2.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Send new config as a new leader.
-    s2.fNet->execReqResp();
-    // Follow-up: commit.
-    s2.fNet->execReqResp();
-    // Wait for bg commit for configuration change.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    CHK_FALSE( s1.raftServer->is_leader() );
-    CHK_TRUE( s2.raftServer->is_leader() );
-    CHK_FALSE( s3.raftServer->is_leader() );
-
-    // Re-yield leadership, now S1 should be the leader again.
-    s2.dbgLog(" --- yield leadership ---");
-    s2.raftServer->yield_leadership();
-    // Send heartbeat.
-    s2.fTimer->invoke( timer_task_type::heartbeat_timer );
-    s2.fNet->execReqResp();
-    // After getting response of heartbeat, S2 will resign.
-    s2.fNet->execReqResp();
-
-    // Now S1 should have received takeover request.
-    // Send vote requests.
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Send new config as a new leader.
-    s1.fNet->execReqResp();
-    // Follow-up: commit.
-    s1.fNet->execReqResp();
-    // Wait for bg commit for configuration change.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    CHK_TRUE( s1.raftServer->is_leader() );
-    CHK_FALSE( s2.raftServer->is_leader() );
-    CHK_FALSE( s3.raftServer->is_leader() );
-
-    print_stats(pkgs);
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-
-    f_base->destroy();
-
-    return 0;
-}
-
-int leadership_takeover_designated_successor_test() {
-    reset_log_files();
-    ptr<FakeNetworkBase> f_base = cs_new<FakeNetworkBase>();
-
-    std::string s1_addr = "S1";
-    std::string s2_addr = "S2";
-    std::string s3_addr = "S3";
-
-    RaftPkg s1(f_base, 1, s1_addr);
-    RaftPkg s2(f_base, 2, s2_addr);
-    RaftPkg s3(f_base, 3, s3_addr);
-    std::vector<RaftPkg*> pkgs = {&s1, &s2, &s3};
-
-    CHK_Z( launch_servers( pkgs ) );
-    CHK_Z( make_group( pkgs ) );
-
-    // Set the priority of S2 to 10.
-    CHK_EQ( raft_server::PrioritySetResult::SET, s1.raftServer->set_priority(2, 10) );
-    // Send priority change reqs.
-    s1.fNet->execReqResp();
-    // Send reqs again for commit.
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Set the priority of S3 to 5.
-    CHK_EQ( raft_server::PrioritySetResult::SET, s1.raftServer->set_priority(3, 5) );
-    // Send priority change reqs.
-    s1.fNet->execReqResp();
-    // Send reqs again for commit.
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Yield leadership to S3.
-    s1.dbgLog(" --- yield leadership ---");
-    s1.raftServer->yield_leadership(false, 3);
-    // Send heartbeat.
-    s1.fTimer->invoke( timer_task_type::heartbeat_timer );
-    s1.fNet->execReqResp();
-    // After getting response of heartbeat, S1 will resign.
-    s1.fNet->execReqResp();
-
-    // Now S3 should have received takeover request.
-    // Send vote requests.
-    s3.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Send new config as a new leader.
-    s3.fNet->execReqResp();
-    // Follow-up: commit.
-    s3.fNet->execReqResp();
-    // Wait for bg commit for configuration change.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    CHK_FALSE( s1.raftServer->is_leader() );
-    CHK_FALSE( s2.raftServer->is_leader() );
-    CHK_TRUE( s3.raftServer->is_leader() );
-
-    // Re-yield leadership to S2.
-    s3.dbgLog(" --- yield leadership ---");
-    s3.raftServer->yield_leadership(false, 2);
-    // Send heartbeat.
-    s3.fTimer->invoke( timer_task_type::heartbeat_timer );
-    s3.fNet->execReqResp();
-    // After getting response of heartbeat, S3 will resign.
-    s3.fNet->execReqResp();
-
-    // Now S2 should have received takeover request.
-    // Send vote requests.
-    s2.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Send new config as a new leader.
-    s2.fNet->execReqResp();
-    // Follow-up: commit.
-    s2.fNet->execReqResp();
-    // Wait for bg commit for configuration change.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    CHK_FALSE( s1.raftServer->is_leader() );
-    CHK_TRUE( s2.raftServer->is_leader() );
-    CHK_FALSE( s3.raftServer->is_leader() );
-
-    // Re-yield leadership with wrong successor,
-    // S1 (highest priority server) will take over.
-    s2.dbgLog(" --- yield leadership ---");
-    s2.raftServer->yield_leadership(false, 12345);
-    // Send heartbeat.
-    s2.fTimer->invoke( timer_task_type::heartbeat_timer );
-    s2.fNet->execReqResp();
-    // After getting response of heartbeat, S2 will resign.
-    s2.fNet->execReqResp();
-
-    // Now S1 should have received takeover request.
-    // Send vote requests.
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Send new config as a new leader.
-    s1.fNet->execReqResp();
-    // Follow-up: commit.
-    s1.fNet->execReqResp();
-    // Wait for bg commit for configuration change.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    CHK_TRUE( s1.raftServer->is_leader() );
-    CHK_FALSE( s2.raftServer->is_leader() );
-    CHK_FALSE( s3.raftServer->is_leader() );
-
-    print_stats(pkgs);
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-
-    f_base->destroy();
-
-    return 0;
-}
-
-int leadership_takeover_by_request_test() {
-    reset_log_files();
-    ptr<FakeNetworkBase> f_base = cs_new<FakeNetworkBase>();
-
-    std::string s1_addr = "S1";
-    std::string s2_addr = "S2";
-    std::string s3_addr = "S3";
-
-    RaftPkg s1(f_base, 1, s1_addr);
-    RaftPkg s2(f_base, 2, s2_addr);
-    RaftPkg s3(f_base, 3, s3_addr);
-    std::vector<RaftPkg*> pkgs = {&s1, &s2, &s3};
-
-    CHK_Z( launch_servers( pkgs ) );
-    CHK_Z( make_group( pkgs ) );
-
-    // Set the priority of S2 to 10.
-    CHK_EQ( raft_server::PrioritySetResult::SET, s1.raftServer->set_priority(2, 10) );
-    // Send priority change reqs.
-    s1.fNet->execReqResp();
-    // Send reqs again for commit.
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Set the priority of S3 to 5.
-    CHK_EQ( raft_server::PrioritySetResult::SET, s1.raftServer->set_priority(3, 5) );
-    // Send priority change reqs.
-    s1.fNet->execReqResp();
-    // Send reqs again for commit.
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Request leadership by the current leader, should fail.
-    CHK_FALSE( s1.raftServer->request_leadership() );
-
-    // Set callback function to refuse resignation.
-    bool refuse_request = true;
-    s1.ctx->set_cb_func([&](cb_func::Type t, cb_func::Param* p) -> cb_func::ReturnCode {
-        if (t != cb_func::Type::ResignationFromLeader) {
-            return cb_default(t, p);
-        }
-        if (refuse_request) {
-            return cb_func::ReturnCode::ReturnNull;
-        }
-        return cb_func::ReturnCode::Ok;
-    });
-
-    // S3 requests the leadership from S1, and it is supposed to be declined.
-    s1.dbgLog(" --- request leadership ---");
-    CHK_TRUE( s3.raftServer->request_leadership() );
-    // Send request.
-    s3.fNet->execReqResp();
-
-    // Send heartbeat.
-    s1.fTimer->invoke( timer_task_type::heartbeat_timer );
-    s1.fNet->execReqResp();
-    s1.fNet->execReqResp();
-
-    // S1 should still be the leader.
-    CHK_TRUE( s1.raftServer->is_leader() );
-
-    // S3 requests the leadership from S1. Now it should succeed.
-    refuse_request = false;
-    s1.dbgLog(" --- request leadership ---");
-    CHK_TRUE( s3.raftServer->request_leadership() );
-    // Send request.
-    s3.fNet->execReqResp();
-
-    // Send heartbeat.
-    s1.fTimer->invoke( timer_task_type::heartbeat_timer );
-    s1.fNet->execReqResp();
-    // After getting response of heartbeat, S1 will resign.
-    s1.fNet->execReqResp();
-
-    // Now S3 should have received takeover request.
-    // Send vote requests.
-    s3.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Send new config as a new leader.
-    s3.fNet->execReqResp();
-    // Follow-up: commit.
-    s3.fNet->execReqResp();
-    // Wait for bg commit for configuration change.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    CHK_FALSE( s1.raftServer->is_leader() );
-    CHK_FALSE( s2.raftServer->is_leader() );
-    CHK_TRUE( s3.raftServer->is_leader() );
-
-    print_stats(pkgs);
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-
-    f_base->destroy();
-
-    return 0;
-}
-
-int leadership_takeover_offline_candidate_test() {
-    reset_log_files();
-    ptr<FakeNetworkBase> f_base = cs_new<FakeNetworkBase>();
-
-    std::string s1_addr = "S1";
-    std::string s2_addr = "S2";
-    std::string s3_addr = "S3";
-
-    RaftPkg s1(f_base, 1, s1_addr);
-    RaftPkg s2(f_base, 2, s2_addr);
-    RaftPkg s3(f_base, 3, s3_addr);
-    std::vector<RaftPkg*> pkgs = {&s1, &s2, &s3};
-
-    raft_params custom_params;
-    custom_params.election_timeout_lower_bound_ = 0;
-    custom_params.election_timeout_upper_bound_ = 1000;
-    custom_params.heart_beat_interval_ = 500;
-    CHK_Z( launch_servers( pkgs, &custom_params ) );
-    CHK_Z( make_group( pkgs ) );
-
-    // Set the priority of S2 to 10.
-    CHK_EQ( raft_server::PrioritySetResult::SET, s1.raftServer->set_priority(2, 10) );
-    // Send priority change reqs.
-    s1.fNet->execReqResp();
-    // Send reqs again for commit.
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Set the priority of S3 to 5.
-    CHK_EQ( raft_server::PrioritySetResult::SET, s1.raftServer->set_priority(3, 5) );
-    // Send priority change reqs.
-    s1.fNet->execReqResp();
-    // Send reqs again for commit.
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Wait longer than heartbeat.
-    TestSuite::sleep_ms(600);
-
-    // Now S2 is offline.
-    s2.fNet->goesOffline();
-
-    // Send heartbeat to S3 only.
-    s1.fTimer->invoke( timer_task_type::heartbeat_timer );
-    s1.fNet->execReqResp();
-
-    // Yield leadership, but now S2 is not responding sudo that
-    // candidate for the next leader should be S3.
-    s1.dbgLog(" --- yield leadership ---");
-    s1.raftServer->yield_leadership();
-
-    // Send heartbeat.
-    s1.fTimer->invoke( timer_task_type::heartbeat_timer );
-    s1.fNet->execReqResp();
-    // After getting response of heartbeat, S1 will resign.
-    s1.fNet->execReqResp();
-
-    // Now S3 should have received takeover request.
-    // Send vote requests.
-    s3.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Send new config as a new leader.
-    s3.fNet->execReqResp();
-    // Follow-up: commit.
-    s3.fNet->execReqResp();
-    // Wait for bg commit for configuration change.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    CHK_FALSE( s1.raftServer->is_leader() );
-    CHK_FALSE( s2.raftServer->is_leader() );
-    CHK_TRUE( s3.raftServer->is_leader() );
-
-    // Re-yield leadership, now S1 should be the leader again.
-    s3.dbgLog(" --- yield leadership ---");
-    s3.raftServer->yield_leadership();
-    // Send heartbeat.
-    s3.fTimer->invoke( timer_task_type::heartbeat_timer );
-    s3.fNet->execReqResp();
-    // After getting response of heartbeat, S2 will resign.
-    s3.fNet->execReqResp();
-
-    // Now S1 should have received takeover request.
-    // Send vote requests.
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Send new config as a new leader.
-    s1.fNet->execReqResp();
-    // Follow-up: commit.
-    s1.fNet->execReqResp();
-    // Wait for bg commit for configuration change.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    CHK_TRUE( s1.raftServer->is_leader() );
-    CHK_FALSE( s2.raftServer->is_leader() );
-    CHK_FALSE( s3.raftServer->is_leader() );
-
-    print_stats(pkgs);
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-
-    f_base->destroy();
-
-    return 0;
-}
-
-int temporary_leader_test() {
-    reset_log_files();
-    ptr<FakeNetworkBase> f_base = cs_new<FakeNetworkBase>();
-
-    std::string s1_addr = "S1";
-    std::string s2_addr = "S2";
-    std::string s3_addr = "S3";
-
-    RaftPkg s1(f_base, 1, s1_addr);
-    RaftPkg s2(f_base, 2, s2_addr);
-    RaftPkg s3(f_base, 3, s3_addr);
-    std::vector<RaftPkg*> pkgs = {&s1, &s2, &s3};
-
-    raft_params custom_params;
-    custom_params.election_timeout_lower_bound_ = 0;
-    custom_params.election_timeout_upper_bound_ = 1000;
-    custom_params.heart_beat_interval_ = 10;
-    CHK_Z( launch_servers( pkgs, &custom_params ) );
-    CHK_Z( make_group( pkgs ) );
-
-    for (auto& entry: pkgs) {
-        RaftPkg* pp = entry;
-        raft_params param = pp->raftServer->get_current_params();
-        param.return_method_ = raft_params::async_handler;
-        pp->raftServer->update_params(param);
-    }
-
-    // Set the priority of S3 to 0.
-    CHK_EQ( raft_server::PrioritySetResult::SET, s1.raftServer->set_priority(3, 0) );
-    // Send priority change reqs.
-    s1.fNet->execReqResp();
-    // Send reqs again for commit.
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Now S2 goes offline.
-    s2.fNet->goesOffline();
-
-    const size_t NUM = 10;
-
-    // Append messages asynchronously.
-    std::list< ptr< cmd_result< ptr<buffer> > > > handlers;
-    for (size_t ii=0; ii<NUM; ++ii) {
-        std::string test_msg = "test" + std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(test_msg.size() + 1);
-        msg->put(test_msg);
-        ptr< cmd_result< ptr<buffer> > > ret =
-            s1.raftServer->append_entries( {msg} );
-
-        CHK_TRUE( ret->get_accepted() );
-
-        handlers.push_back(ret);
-    }
-
-    // Replicate.
-    for (size_t ii=0; ii<3; ++ii) s1.fNet->execReqResp();
-    // Wait for bg commit.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // All handlers should be OK.
-    for (auto& entry: handlers) {
-        CHK_TRUE( entry->has_result() );
-        CHK_EQ( cmd_result_code::OK, entry->get_result_code() );
-    }
-
-    // Now S1 goes offline, and S2 goes online.
-    s1.fNet->goesOffline();
-    s2.fNet->goesOnline();
-
-    const size_t MAX_ATTEMPTS = 100;
-    size_t attempts = 0;
-    do {
-        // Vote, S2 should be rejected all the time.
-        s2.fTimer->invoke( timer_task_type::election_timer );
-        // Pre-vote and vote.
-        s2.fNet->execReqResp();
-        s2.fNet->execReqResp();
-
-        s3.fTimer->invoke( timer_task_type::election_timer );
-        // Pre-vote and vote.
-        s3.fNet->execReqResp();
-        s3.fNet->execReqResp();
-
-        // Sleep.
-        attempts++;
-        TestSuite::sleep_ms(custom_params.heart_beat_interval_);
-
-        // Repeat until S3 becomes (temporary) leader.
-    } while (!s3.raftServer->is_leader() && attempts < MAX_ATTEMPTS);
-    CHK_SM(attempts, MAX_ATTEMPTS);
-
-    // Commit for reconfigure.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Now S3 will yield leadership for S2.
-    s3.fTimer->invoke( timer_task_type::heartbeat_timer );
-    // Catch-up and commit.
-    s3.fNet->execReqResp();
-    s3.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Resign.
-    s3.fNet->execReqResp();
-    s3.fNet->execReqResp();
-
-    // Now S2 initiates leader election.
-    s2.fNet->execReqResp();
-    s2.fNet->execReqResp();
-
-    CHK_TRUE( s2.raftServer->is_leader() );
-    s2.fNet->execReqResp();
-    s2.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    print_stats(pkgs);
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-
-    f_base->destroy();
-
-    return 0;
-}
-
 int check_priorities(const std::vector<RaftPkg*>& pkgs,
                      const std::vector<int>& priorities) {
     for (auto& entry: pkgs)
@@ -1991,637 +1047,6 @@ int follower_reconnect_test() {
     fake_executer_killer(&exec_args);
     hh.join();
     CHK_Z( hh.getResult() );
-
-    f_base->destroy();
-
-    return 0;
-}
-
-int snapshot_basic_test() {
-    reset_log_files();
-    ptr<FakeNetworkBase> f_base = cs_new<FakeNetworkBase>();
-
-    std::string s1_addr = "S1";
-    std::string s2_addr = "S2";
-    std::string s3_addr = "S3";
-
-    RaftPkg s1(f_base, 1, s1_addr);
-    RaftPkg s2(f_base, 2, s2_addr);
-    RaftPkg s3(f_base, 3, s3_addr);
-    std::vector<RaftPkg*> pkgs = {&s1, &s2, &s3};
-
-    CHK_Z( launch_servers( pkgs ) );
-    CHK_Z( make_group( pkgs ) );
-
-    // Append a message using separate thread.
-    ExecArgs exec_args(&s1);
-    TestSuite::ThreadHolder hh(&exec_args, fake_executer, fake_executer_killer);
-
-    for (size_t ii=0; ii<5; ++ii) {
-        std::string test_msg = "test" + std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(test_msg.size() + 1);
-        msg->put(test_msg);
-        exec_args.setMsg(msg);
-        exec_args.eaExecuter.invoke();
-
-        // Wait for executer thread.
-        TestSuite::sleep_ms(EXECUTOR_WAIT_MS);
-
-        CHK_NULL( exec_args.getMsg().get() );
-
-        // NOTE: Send it to S2 only, S3 will be lagging behind.
-        s1.fNet->execReqResp("S2"); // replication.
-        s1.fNet->execReqResp("S2"); // commit.
-        CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) ); // commit execution.
-    }
-    // Make req to S3 failed.
-    s1.fNet->makeReqFail("S3");
-
-    // Trigger heartbeat to S3, it will initiate snapshot transmission.
-    s1.fTimer->invoke(timer_task_type::heartbeat_timer);
-    s1.fNet->execReqResp();
-
-    // Send the entire snapshot.
-    do {
-        s1.fNet->execReqResp();
-    } while (s3.raftServer->is_receiving_snapshot());
-
-    s1.fNet->execReqResp(); // commit.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) ); // commit execution.
-
-    // State machine should be identical.
-    CHK_OK( s2.getTestSm()->isSame( *s1.getTestSm() ) );
-    CHK_OK( s3.getTestSm()->isSame( *s1.getTestSm() ) );
-
-    // There shouldn't be any open snapshot ctx.
-    CHK_Z( s1.getTestSm()->getNumOpenedUserCtxs() );
-
-    print_stats(pkgs);
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-
-    fake_executer_killer(&exec_args);
-    hh.join();
-    CHK_Z( hh.getResult() );
-
-    f_base->destroy();
-
-    return 0;
-}
-
-int snapshot_new_member_restart_test() {
-    reset_log_files();
-    ptr<FakeNetworkBase> f_base = cs_new<FakeNetworkBase>();
-
-    std::string s1_addr = "S1";
-    std::string s2_addr = "S2";
-    std::string s3_addr = "S3";
-
-    RaftPkg s1(f_base, 1, s1_addr);
-    RaftPkg s2(f_base, 2, s2_addr);
-    RaftPkg s3(f_base, 3, s3_addr);
-    std::vector<RaftPkg*> pkgs = {&s1, &s2, &s3};
-    std::vector<RaftPkg*> pkgs_orig = {&s1, &s2};
-
-    CHK_Z( launch_servers( pkgs ) );
-    CHK_Z( make_group( pkgs_orig ) );
-
-    // Append a message using separate thread.
-    ExecArgs exec_args(&s1);
-    TestSuite::ThreadHolder hh(&exec_args, fake_executer, fake_executer_killer);
-
-    for (size_t ii=0; ii<5; ++ii) {
-        std::string test_msg = "test" + std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(test_msg.size() + 1);
-        msg->put(test_msg);
-        exec_args.setMsg(msg);
-        exec_args.eaExecuter.invoke();
-
-        // Wait for executer thread.
-        TestSuite::sleep_ms(EXECUTOR_WAIT_MS);
-
-        CHK_NULL( exec_args.getMsg().get() );
-
-        s1.fNet->execReqResp(); // replication.
-        s1.fNet->execReqResp(); // commit.
-        CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) ); // commit execution.
-    }
-
-    // Add S3 to S1.
-    s1.raftServer->add_srv(*(s3.getTestMgr()->get_srv_config()));
-
-    // Join req/resp.
-    s1.fNet->execReqResp();
-    // Add new server, notify existing peers.
-    // After getting response, it will make configuration commit.
-    s1.fNet->execReqResp();
-    // Notify new commit, start snapshot transmission.
-    s1.fNet->execReqResp();
-    // Wait for bg commit for configuration change.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Shutdown S3.
-    s1.dbgLog(" --- shutting down S3 ---");
-    s3.raftServer->shutdown();
-    s3.fNet->shutdown();
-    s1.dbgLog(" --- shut down S3 ---");
-
-    // Trigger heartbeat, to close connection.
-    s1.fTimer->invoke(timer_task_type::heartbeat_timer);
-    s1.fNet->execReqResp();
-
-    // Restart s3.
-    s1.dbgLog(" --- restarting S3 ---");
-    CHK_Z( launch_servers( {&s3},
-                           /* custom_params = */ nullptr,
-                           /* restart = */ true ) );
-    s1.dbgLog(" --- restarted S3 ---");
-
-    // Trigger heartbeat, to resume snapshot transmission.
-    s1.fTimer->invoke(timer_task_type::heartbeat_timer);
-    s1.fNet->execReqResp();
-    s1.fNet->execReqResp();
-
-    // Send the entire snapshot.
-    do {
-        s1.fNet->execReqResp();
-    } while (s3.raftServer->is_receiving_snapshot());
-
-    // commit.
-    s1.fNet->execReqResp();
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) ); // commit execution.
-
-    // One more heartbeat.
-    s1.fTimer->invoke(timer_task_type::heartbeat_timer);
-    s1.fNet->execReqResp();
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) ); // commit execution.
-
-    // State machine should be identical.
-    CHK_OK( s2.getTestSm()->isSame( *s1.getTestSm() ) );
-    CHK_OK( s3.getTestSm()->isSame( *s1.getTestSm() ) );
-
-    print_stats(pkgs);
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-
-    fake_executer_killer(&exec_args);
-    hh.join();
-    CHK_Z( hh.getResult() );
-
-    f_base->destroy();
-
-    return 0;
-}
-
-int snapshot_manual_creation_test() {
-    reset_log_files();
-    ptr<FakeNetworkBase> f_base = cs_new<FakeNetworkBase>();
-
-    std::string s1_addr = "S1";
-    std::string s2_addr = "S2";
-    std::string s3_addr = "S3";
-
-    RaftPkg s1(f_base, 1, s1_addr);
-    RaftPkg s2(f_base, 2, s2_addr);
-    RaftPkg s3(f_base, 3, s3_addr);
-    std::vector<RaftPkg*> pkgs = {&s1, &s2, &s3};
-
-    CHK_Z( launch_servers( pkgs ) );
-    CHK_Z( make_group( pkgs ) );
-
-    // Append a message using separate thread.
-    ExecArgs exec_args(&s1);
-    TestSuite::ThreadHolder hh(&exec_args, fake_executer, fake_executer_killer);
-
-    for (auto& entry: pkgs) {
-        RaftPkg* pp = entry;
-        raft_params param = pp->raftServer->get_current_params();
-        param.return_method_ = raft_params::async_handler;
-        param.snapshot_distance_ = 100;
-        pp->raftServer->update_params(param);
-    }
-
-    const size_t NUM = 10;
-
-    // Append messages asynchronously.
-    std::list< ptr< cmd_result< ptr<buffer> > > > handlers;
-    for (size_t ii=0; ii<NUM; ++ii) {
-        std::string test_msg = "test" + std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(test_msg.size() + 1);
-        msg->put(test_msg);
-        ptr< cmd_result< ptr<buffer> > > ret =
-            s1.raftServer->append_entries( {msg} );
-
-        CHK_TRUE( ret->get_accepted() );
-
-        handlers.push_back(ret);
-    }
-
-    // NOTE: Send it to S2 only, S3 will be lagging behind.
-    s1.fNet->execReqResp("S2"); // replication.
-    s1.fNet->execReqResp("S2"); // commit.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) ); // commit execution.
-
-    // One more time to make sure.
-    s1.fNet->execReqResp("S2");
-    s1.fNet->execReqResp("S2");
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Remember the current commit index.
-    uint64_t committed_index = s1.raftServer->get_committed_log_idx();
-
-    // Create a manual snapshot.
-    ulong log_idx = s1.raftServer->create_snapshot();
-    CHK_EQ( committed_index, log_idx );
-    CHK_EQ( log_idx, s1.raftServer->get_last_snapshot_idx() );
-
-    // Make req to S3 failed.
-    s1.fNet->makeReqFail("S3");
-
-    // Trigger heartbeat to S3, it will initiate snapshot transmission.
-    s1.fTimer->invoke(timer_task_type::heartbeat_timer);
-    s1.fNet->execReqResp();
-
-    // Send the entire snapshot.
-    do {
-        s1.fNet->execReqResp();
-    } while (s3.raftServer->is_receiving_snapshot());
-
-    s1.fNet->execReqResp(); // commit.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) ); // commit execution.
-
-    // State machine should be identical.
-    CHK_OK( s2.getTestSm()->isSame( *s1.getTestSm() ) );
-    CHK_OK( s3.getTestSm()->isSame( *s1.getTestSm() ) );
-
-    CHK_EQ( committed_index, s3.getTestSm()->last_snapshot()->get_last_log_idx() );
-
-    // There shouldn't be any open snapshot ctx.
-    CHK_Z( s1.getTestSm()->getNumOpenedUserCtxs() );
-
-    print_stats(pkgs);
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-
-    fake_executer_killer(&exec_args);
-    hh.join();
-    CHK_Z( hh.getResult() );
-
-    f_base->destroy();
-
-    return 0;
-}
-
-int snapshot_creation_index_inversion_test() {
-    reset_log_files();
-    ptr<FakeNetworkBase> f_base = cs_new<FakeNetworkBase>();
-
-    std::string s1_addr = "S1";
-    std::string s2_addr = "S2";
-    std::string s3_addr = "S3";
-
-    RaftPkg s1(f_base, 1, s1_addr);
-    RaftPkg s2(f_base, 2, s2_addr);
-    RaftPkg s3(f_base, 3, s3_addr);
-    std::vector<RaftPkg*> pkgs = {&s1, &s2, &s3};
-
-    CHK_Z( launch_servers( pkgs ) );
-    CHK_Z( make_group( pkgs ) );
-
-    // Append a message using separate thread.
-    ExecArgs exec_args(&s1);
-    TestSuite::ThreadHolder hh(&exec_args, fake_executer, fake_executer_killer);
-
-    for (auto& entry: pkgs) {
-        RaftPkg* pp = entry;
-        raft_params param = pp->raftServer->get_current_params();
-        param.return_method_ = raft_params::async_handler;
-        pp->raftServer->update_params(param);
-    }
-
-    const size_t NUM = 5;
-
-    // Set a callback function to manually create snapshot,
-    // right before the automatic snapshot creation.
-    bool manual_snp_creation_succ = false;
-    s1.ctx->set_cb_func([&](cb_func::Type t, cb_func::Param* p) -> cb_func::ReturnCode {
-        // At the beginning of an automatic snapshot creation,
-        // create a manual snapshot, to mimic index inversion.
-        if (t != cb_func::Type::SnapshotCreationBegin) {
-            return cb_default(t, p);
-        }
-
-        // This function should be invoked only once, to avoid
-        // infinite recursive call.
-        static bool invoked = false;
-        if (!invoked) {
-            invoked = true;
-            ulong log_idx = s1.raftServer->create_snapshot();
-            manual_snp_creation_succ = (log_idx > 0);
-        }
-        return cb_func::ReturnCode::Ok;
-    });
-
-    // Append messages asynchronously.
-    std::list< ptr< cmd_result< ptr<buffer> > > > handlers;
-    for (size_t ii=0; ii<NUM; ++ii) {
-        std::string test_msg = "test" + std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(test_msg.size() + 1);
-        msg->put(test_msg);
-        ptr< cmd_result< ptr<buffer> > > ret =
-            s1.raftServer->append_entries( {msg} );
-
-        CHK_TRUE( ret->get_accepted() );
-
-        handlers.push_back(ret);
-    }
-
-    // NOTE: Send it to S2 only, S3 will be lagging behind.
-    s1.fNet->execReqResp("S2"); // replication.
-    s1.fNet->execReqResp("S2"); // commit.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) ); // commit execution.
-
-    // One more time to make sure.
-    s1.fNet->execReqResp("S2");
-    s1.fNet->execReqResp("S2");
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Snapshot creation should have happened only once, by manual creation.
-    CHK_TRUE(manual_snp_creation_succ);
-    CHK_EQ(1, s1.getTestSm()->getNumSnapshotCreations());
-
-    print_stats(pkgs);
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-
-    fake_executer_killer(&exec_args);
-    hh.join();
-    CHK_Z( hh.getResult() );
-
-    f_base->destroy();
-
-    return 0;
-}
-
-int snapshot_scheduled_creation_test() {
-    reset_log_files();
-    ptr<FakeNetworkBase> f_base = cs_new<FakeNetworkBase>();
-
-    std::string s1_addr = "S1";
-    std::string s2_addr = "S2";
-    std::string s3_addr = "S3";
-
-    RaftPkg s1(f_base, 1, s1_addr);
-    RaftPkg s2(f_base, 2, s2_addr);
-    RaftPkg s3(f_base, 3, s3_addr);
-    std::vector<RaftPkg*> pkgs = {&s1, &s2, &s3};
-
-    CHK_Z( launch_servers( pkgs ) );
-    CHK_Z( make_group( pkgs ) );
-
-    // Append a message using separate thread.
-    ExecArgs exec_args(&s1);
-    TestSuite::ThreadHolder hh(&exec_args, fake_executer, fake_executer_killer);
-
-    for (auto& entry: pkgs) {
-        RaftPkg* pp = entry;
-        raft_params param = pp->raftServer->get_current_params();
-        param.return_method_ = raft_params::async_handler;
-        pp->raftServer->update_params(param);
-    }
-
-    const size_t NUM = 5;
-
-    // Append messages asynchronously.
-    std::list< ptr< cmd_result< ptr<buffer> > > > handlers;
-    for (size_t ii = 0; ii < NUM; ++ii) {
-        std::string test_msg = "test" + std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(test_msg.size() + 1);
-        msg->put(test_msg);
-        ptr< cmd_result< ptr<buffer> > > ret =
-            s1.raftServer->append_entries( {msg} );
-
-        CHK_TRUE( ret->get_accepted() );
-
-        handlers.push_back(ret);
-    }
-
-    s1.fNet->execReqResp(); // replication.
-    s1.fNet->execReqResp(); // commit.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) ); // commit execution.
-
-    // One more time to make sure.
-    s1.fNet->execReqResp();
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Manually create a snapshot.
-    uint64_t log_idx = s1.raftServer->create_snapshot();
-    CHK_GT(log_idx, 0);
-
-    // Schedule snapshot creation and wait 500ms, there shouldn't be any progress.
-    auto sched_ret = s1.raftServer->schedule_snapshot_creation();
-    TestSuite::sleep_ms(500, "wait for async snapshot creation");
-    CHK_FALSE(sched_ret->has_result());
-
-    uint64_t last_idx = s1.raftServer->get_last_log_idx();
-
-    // Append more messages asynchronously.
-    for (size_t ii = NUM; ii < NUM * 2; ++ii) {
-        std::string test_msg = "test" + std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(test_msg.size() + 1);
-        msg->put(test_msg);
-        ptr< cmd_result< ptr<buffer> > > ret =
-            s1.raftServer->append_entries( {msg} );
-
-        CHK_TRUE( ret->get_accepted() );
-
-        handlers.push_back(ret);
-    }
-
-    s1.fNet->execReqResp(); // replication.
-    s1.fNet->execReqResp(); // commit.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) ); // commit execution.
-
-    // One more time to make sure.
-    s1.fNet->execReqResp();
-    s1.fNet->execReqResp();
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Now it should have the result.
-    CHK_TRUE(sched_ret->has_result());
-    CHK_EQ(last_idx + 1, sched_ret->get());
-
-    print_stats(pkgs);
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-
-    fake_executer_killer(&exec_args);
-    hh.join();
-    CHK_Z( hh.getResult() );
-
-    f_base->destroy();
-
-    return 0;
-}
-
-int snapshot_randomized_creation_test() {
-    reset_log_files();
-    ptr<FakeNetworkBase> f_base = cs_new<FakeNetworkBase>();
-
-    std::string s1_addr = "S1";
-    std::string s2_addr = "S2";
-    std::string s3_addr = "S3";
-
-    RaftPkg s1(f_base, 1, s1_addr);
-    RaftPkg s2(f_base, 2, s2_addr);
-    RaftPkg s3(f_base, 3, s3_addr);
-    std::vector<RaftPkg*> pkgs = {&s1, &s2, &s3};
-
-    const size_t NUM = 50;
-
-    raft_params params;
-    params.with_randomized_snapshot_creation_enabled(true);
-    params.with_election_timeout_lower(0);
-    params.with_election_timeout_upper(10000);
-    params.with_hb_interval(5000);
-    params.with_client_req_timeout(1000000);
-    params.with_reserved_log_items(0);
-    params.with_snapshot_enabled(NUM);
-    params.with_log_sync_stopping_gap(1);
-
-    CHK_Z( launch_servers( pkgs, &params ) );
-    CHK_Z( make_group( pkgs ) );
-
-    // Append a message using separate thread.
-    ExecArgs exec_args(&s1);
-    TestSuite::ThreadHolder hh(&exec_args, fake_executer, fake_executer_killer);
-
-    for (auto& entry: pkgs) {
-        RaftPkg* pp = entry;
-        raft_params param = pp->raftServer->get_current_params();
-        param.return_method_ = raft_params::async_handler;
-        pp->raftServer->update_params(param);
-    }
-
-    // Append messages asynchronously.
-    std::list< ptr< cmd_result< ptr<buffer> > > > handlers;
-    for (size_t ii=0; ii<NUM; ++ii) {
-        std::string test_msg = "test" + std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(test_msg.size() + 1);
-        msg->put(test_msg);
-        ptr< cmd_result< ptr<buffer> > > ret =
-            s1.raftServer->append_entries( {msg} );
-
-        CHK_TRUE( ret->get_accepted() );
-
-        handlers.push_back(ret);
-    }
-
-    // NOTE: Send it to S2, S3
-    s1.fNet->execReqResp(); // replication.
-    s1.fNet->execReqResp(); // commit.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) ); // commit execution.
-
-    // One more time to make sure.
-    s1.fNet->execReqResp(); // replication.
-    s1.fNet->execReqResp(); // commit.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) ); // commit execution.
-
-    CHK_NEQ(NUM, s1.getTestSm()->last_snapshot()->get_last_log_idx())
-    CHK_NEQ(NUM, s2.getTestSm()->last_snapshot()->get_last_log_idx())
-    CHK_NEQ(NUM, s3.getTestSm()->last_snapshot()->get_last_log_idx())
-
-    print_stats(pkgs);
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
-    s3.raftServer->shutdown();
-
-    fake_executer_killer(&exec_args);
-    hh.join();
-    CHK_Z( hh.getResult() );
-
-    f_base->destroy();
-
-    return 0;
-}
-
-int snapshot_close_for_removed_peer_test() {
-    reset_log_files();
-    ptr<FakeNetworkBase> f_base = cs_new<FakeNetworkBase>();
-
-    std::string s1_addr = "S1";
-    std::string s2_addr = "S2";
-
-    RaftPkg s1(f_base, 1, s1_addr);
-    RaftPkg s2(f_base, 2, s2_addr);
-    std::vector<RaftPkg*> pkgs = {&s1, &s2};
-
-    CHK_Z( launch_servers( pkgs ) );
-    CHK_Z( make_group( pkgs ) );
-
-    for (auto& entry: pkgs) {
-        RaftPkg* pp = entry;
-        raft_params param = pp->raftServer->get_current_params();
-        param.return_method_ = raft_params::async_handler;
-        // Set quorum to 1 so as to make S1 commits data locally.
-        param.custom_commit_quorum_size_ = 1;
-        param.custom_election_quorum_size_ = 1;
-        pp->raftServer->update_params(param);
-    }
-
-    const size_t NUM = 10;
-
-    // Append messages asynchronously.
-    std::list< ptr< cmd_result< ptr<buffer> > > > handlers;
-    for (size_t ii=0; ii<NUM; ++ii) {
-        std::string test_msg = "test" + std::to_string(ii);
-        ptr<buffer> msg = buffer::alloc(test_msg.size() + 1);
-        msg->put(test_msg);
-        ptr< cmd_result< ptr<buffer> > > ret =
-            s1.raftServer->append_entries( {msg} );
-
-        CHK_TRUE( ret->get_accepted() );
-
-        handlers.push_back(ret);
-    }
-    // Wait for bg commit.
-    CHK_Z( wait_for_sm_exec(pkgs, COMMIT_TIMEOUT_SEC) );
-
-    // Make req to S2 failed.
-    s1.fNet->makeReqFailAll("S2");
-
-    // Heartbeat, this will initiate snapshot transfer to S2.
-    s1.fTimer->invoke(timer_task_type::heartbeat_timer);
-    s1.fNet->execReqResp("S2");
-
-    // Now remove S2.
-    s1.raftServer->remove_srv(2);
-
-    // Heartbeat, and make request fail.
-    s1.fTimer->invoke(timer_task_type::heartbeat_timer);
-    s1.fNet->makeReqFailAll("S2");
-
-    // After S2 is removed, the snapshot ctx should be destroyed.
-    CHK_Z( s1.getTestSm()->getNumOpenedUserCtxs() );
-
-    print_stats(pkgs);
-
-    s1.raftServer->shutdown();
-    s2.raftServer->shutdown();
 
     f_base->destroy();
 
@@ -3578,33 +2003,6 @@ int main(int argc, char** argv) {
     ts.doTest( "multiple config change test",
                multiple_config_change_test );
 
-    ts.doTest( "leader election basic test",
-               leader_election_basic_test );
-
-    ts.doTest( "leader election priority test",
-               leader_election_priority_test );
-
-    ts.doTest( "leader election with aggressive node test",
-               leader_election_with_aggressive_node_test );
-
-    ts.doTest( "leader election with catching-up server test",
-               leader_election_with_catching_up_server_test );
-
-    ts.doTest( "leadership takeover basic test",
-               leadership_takeover_basic_test );
-
-    ts.doTest( "leadership takeover with designated successor test",
-               leadership_takeover_designated_successor_test );
-
-    ts.doTest( "leadership takeover by request test",
-               leadership_takeover_by_request_test );
-
-    ts.doTest( "leadership takeover with offline candidate test",
-               leadership_takeover_offline_candidate_test );
-
-    ts.doTest( "temporary leader test",
-               temporary_leader_test );
-
     ts.doTest( "priority broadcast test",
                priority_broadcast_test );
 
@@ -3616,27 +2014,6 @@ int main(int argc, char** argv) {
 
     ts.doTest( "follower reconnect test",
                follower_reconnect_test );
-
-    ts.doTest( "snapshot basic test",
-               snapshot_basic_test );
-
-    ts.doTest( "snapshot new member restart test",
-               snapshot_new_member_restart_test );
-
-    ts.doTest( "snapshot manual creation test",
-               snapshot_manual_creation_test );
-
-    ts.doTest( "snapshot creation index inversion test",
-               snapshot_creation_index_inversion_test );
-
-    ts.doTest( "snapshot scheduled creation test",
-               snapshot_scheduled_creation_test );
-
-    ts.doTest( "snapshot randomized creation test",
-               snapshot_randomized_creation_test );
-
-    ts.doTest( "snapshot close for removed peer test",
-               snapshot_close_for_removed_peer_test );
 
     ts.doTest( "join empty node test",
                join_empty_node_test );
